Split value visitor parts into separate mods with better encapsulation
