+++ crates/lib/runner-state/src/lib.rs
@@ -1,5 +1,14 @@
 mod state;
 mod util;
-pub mod value_visitor;
+mod value {
+    pub mod evaluator;
+    pub mod resolver;
+    pub mod source_collector;
+}
 
 pub use self::state::*;
+
+use self::value::resolver::resolve_value_tree;
+use self::value::source_collector::collect_value_sources;
+
+pub use self::value::evaluator::ValueExprEvaluator;
